// src/types/workDuration.ts
export interface IWorkDuration {
    totalWorkDuration: string;
    averageTaskDuration: string;
    netWorkTime: string;
    latencyTime: string;
    codingTime: string;
    codeReviewTime: string;
    qaTime: string;
    idleTime: string;
    averageDelegationTime: string;
    percentageTimeWithoutWork: string;
    taskVelocity: string;
    workVelocity: string;
}